BUILD WORKING on osx10.8 xcode4.6.2 Had to remove NSTextFinder.h & move the enum to a relevant place, as well as a bunch of other fixes

+++ BufferUtils.c
@@ -411,7 +411,7 @@ CFStringRef CreateRandomizedFileName() {
     }
     
     CFStringRef name = CFStringCreateWithFormat(kCFAllocatorDefault, NULL, CFSTR(".%lu%lu-%d-%d"), 
-						psn.highLongOfPSN, psn.lowLongOfPSN, (int)CFAbsoluteTimeGetCurrent(), sequence);
+						psn.highLongOfPSN, (NSUInteger)psn.lowLongOfPSN, (int)CFAbsoluteTimeGetCurrent(), sequence);
     
     return name;
 }
